+++ mobile/Features/ContentLibrary/Components/FolderCreatedModal.jsx
@@ -1,12 +1,12 @@
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 import {
-  View, Text, Modal, TouchableOpacity, StyleSheet, Image, ScrollView, Pressable,
-  TextInput, TouchableWithoutFeedback, Keyboard,
+  View, Text, Modal, Image, Pressable, TouchableWithoutFeedback, Keyboard,
 } from 'react-native';
 
 import PropTypes from 'prop-types';
 import exit from '../../../assets/exit.png';
 import check from '../../../assets/green_check.png';
+import styles from './FolderModalStyle';
 
 function FolderCreatedModal({ modalVisibleParent, toggleModal, newFolderName }) {
   const handleOK = () => {
@@ -55,7 +55,7 @@ function FolderCreatedModal({ modalVisibleParent, toggleModal, newFolderName }) 
                 The resource is now in your folder
               </Text>
               <Text style={{
-                textDecorationLine: 'underline', ontFamily: 'cabinet-grotesk-regular', fontSize: 18, color: '#676C6C', paddingBottom: 15,
+                textDecorationLine: 'underline', fontFamily: 'cabinet-grotesk-regular', fontSize: 18, color: '#676C6C', paddingBottom: 15,
               }}
               >
                 {' '}
@@ -78,73 +78,6 @@ function FolderCreatedModal({ modalVisibleParent, toggleModal, newFolderName }) 
     </Modal>
   );
 }
-
-const styles = StyleSheet.create({
-  centeredView: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: 'rgba(0,0,0,0.5)', // semi-transparent background
-  },
-  modalView: {
-    backgroundColor: '#DFE5E5',
-    borderRadius: 10,
-    padding: 20,
-    width: '80%',
-    borderWidth: 1,
-    borderColor: 'black',
-
-  },
-  inputFolder: {
-    height: 33,
-    width: '100%',
-    borderColor: 'gray',
-    borderWidth: 1,
-    marginBottom: 20,
-    paddingHorizontal: 10,
-    backgroundColor: '#C6CECE',
-    borderRadius: 8,
-
-  },
-  inputDescription: {
-    height: 48,
-    width: '100%',
-    borderColor: 'gray',
-    borderWidth: 1,
-    marginBottom: 20,
-    paddingHorizontal: 10,
-    backgroundColor: '#C6CECE',
-    borderRadius: 8,
-
-  },
-  saveButton: {
-    height: 40,
-    width: '80%',
-    borderColor: 'gray',
-    borderWidth: 1,
-    marginBottom: 20,
-    paddingHorizontal: 10,
-    borderRadius: 99,
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: '#374342',
-
-  },
-  linearGradient: {
-    flex: 1,
-    paddingLeft: 15,
-    paddingRight: 15,
-    borderRadius: 5,
-  },
-  buttonText: {
-    fontSize: 18,
-    fontFamily: 'Gill Sans',
-    textAlign: 'center',
-    margin: 10,
-    color: '#ffffff',
-    backgroundColor: 'transparent',
-  },
-});
 
 export default FolderCreatedModal;
