+++ mobile/Features/ContentLibrary/Components/FolderModalStyle.jsx
@@ -1,0 +1,70 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  centeredView: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+    backgroundColor: 'rgba(0,0,0,0.5)', // semi-transparent background
+  },
+  modalView: {
+    backgroundColor: '#DFE5E5',
+    borderRadius: 10,
+    padding: 20,
+    width: '80%',
+    borderWidth: 1,
+    borderColor: 'black',
+
+  },
+  inputFolder: {
+    height: 33,
+    width: '100%',
+    borderColor: 'gray',
+    borderWidth: 1,
+    marginBottom: 20,
+    paddingHorizontal: 10,
+    backgroundColor: '#C6CECE',
+    borderRadius: 8,
+
+  },
+  inputDescription: {
+    height: 48,
+    width: '100%',
+    borderColor: 'gray',
+    borderWidth: 1,
+    marginBottom: 20,
+    paddingHorizontal: 10,
+    backgroundColor: '#C6CECE',
+    borderRadius: 8,
+
+  },
+  saveButton: {
+    height: 40,
+    width: '80%',
+    borderColor: 'gray',
+    borderWidth: 1,
+    marginBottom: 20,
+    paddingHorizontal: 10,
+    borderRadius: 99,
+    justifyContent: 'center',
+    alignItems: 'center',
+    backgroundColor: '#374342',
+
+  },
+  linearGradient: {
+    flex: 1,
+    paddingLeft: 15,
+    paddingRight: 15,
+    borderRadius: 5,
+  },
+  buttonText: {
+    fontSize: 18,
+    fontFamily: 'Gill Sans',
+    textAlign: 'center',
+    margin: 10,
+    color: '#ffffff',
+    backgroundColor: 'transparent',
+  },
+});
+
+export default styles;
